+++ index.html
@@ -236,7 +236,7 @@
     </section>
   </template>
 </body>
-<script src="js/functions.js"></script>
-<script src="js/main.js"></script>
+<script src="js/functions.js" type="modele"></script>
+<script src="js/main.js" type="module"></script>
 
 </html>
